backend/requirements.txt
Django>=5.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
python-decouple==3.8
django-extensions==3.2.3
werkzeug==3.0.1
pyOpenSSL>=23.0.0

